+++ CHANGELOG.md
@@ -8,6 +8,15 @@ Currently, this repo is in Prerelease. When it is released, this project will ad
 
 ## Prerelease
 
+### Updates
+
+- Updates `Notification` variant font colors to sync with VDL.
+- Updates `Tooltip` text color.
+- Updates the `SearchBar` component to remove the text input right border.
+- Updates the `SkeletonLoader` component to make the button placeholder look more like a button.
+
+## 3.6.0 (April 10, 2025)
+
 ### Adds
 
 - Adds the `"navigationSubdirectoryArrowLeft"` and `"navigationSubdirectoryArrowRight"` options to the `Icon` component.